requirements.txt
openai-whisper
PyQt5
pyaudio
numpy 